Update Dockerfile
LIGO hosts Jupyterlab instances which are really useful. These run on underlying kernels, and the PyCBC releases should be one of those. Running these kernels from this Docker image (or using a conda/CVMFS build) seems to be the preferred solution. Adding in two additional dependencies to facilitate this as requested. (I'll check with with Philippe Grassia who requested this before wanting to merge).